Moved MidiSetup into midi package
