+++ src/main/java/net/perkowitz/issho/controller/apps/draw/Draw.java
@@ -4,7 +4,7 @@ import com.google.common.collect.Lists;
 import net.perkowitz.issho.controller.Colors;
 import net.perkowitz.issho.controller.Controller;
 import net.perkowitz.issho.controller.ControllerListener;
-import net.perkowitz.issho.controller.MidiSetup;
+import net.perkowitz.issho.controller.midi.MidiSetup;
 import net.perkowitz.issho.controller.novation.LaunchpadPro;
 import net.perkowitz.issho.controller.yaeltex.YaeltexHachiXL;
