+++ src/main/java/net/perkowitz/issho/controller/apps/viz/Viz.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 import net.perkowitz.issho.controller.Colors;
 import net.perkowitz.issho.controller.Controller;
 import net.perkowitz.issho.controller.ControllerListener;
-import net.perkowitz.issho.controller.MidiSetup;
+import net.perkowitz.issho.controller.midi.MidiSetup;
 import net.perkowitz.issho.controller.novation.LaunchpadPro;
 import net.perkowitz.issho.controller.yaeltex.YaeltexHachiXL;
