+++ src/main/java/net/perkowitz/issho/controller/yaeltex/Palette.java
@@ -1,16 +1,9 @@
 // Test runs a simple drawing program to test the Yaeltex controller.
 package net.perkowitz.issho.controller.yaeltex;
 
-import net.perkowitz.issho.controller.Colors;
 import net.perkowitz.issho.controller.Controller;
-import net.perkowitz.issho.controller.MidiSetup;
-import net.perkowitz.issho.controller.elements.Button;
-import net.perkowitz.issho.controller.elements.Knob;
+import net.perkowitz.issho.controller.midi.MidiSetup;
 import net.perkowitz.issho.controller.elements.Pad;
-
-import java.awt.*;
-
-import static net.perkowitz.issho.controller.Colors.BLACK;
 
 /**
  * Created by mikep on 7/28/20.
